adding user-profile

+++ frontend/src/app/group-list/group-list.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import {Group, GroupService} from '../services/group.service';
+import {User, UserService} from '../services/user.service';
 import {HttpClient} from '@angular/common/http';
 
 
@@ -12,9 +13,9 @@ import {HttpClient} from '@angular/common/http';
 export class GroupListComponent implements OnInit {
 
   groups: Group[];
-  displayedColumns = ['name', 'edit', 'delete'];
+  displayedColumns = ['name', 'leader', 'edit', 'delete'];
 
-  constructor(private http: HttpClient, private groupService: GroupService) { }
+  constructor(private http: HttpClient, private groupService: GroupService, private userService: UserService) { }
 
   ngOnInit(): void {
     this.retrieveGroups();
@@ -28,11 +29,16 @@ export class GroupListComponent implements OnInit {
     });
   }
 
-  deleteEvent(group: Group): void {
-    this.groupService.deleteGroup(group)
-      .subscribe(() => {
-        this.retrieveGroups();
-        alert('deleted successfully!');
-      });
+  deleteGroup(group: Group): void {
+    if (localStorage.getItem('currentUser') == group.leader || localStorage.getItem('currentUser') == "admin") {
+      this.groupService.deleteGroup(group)
+        .subscribe(() => {
+          this.retrieveGroups();
+          alert('deleted successfully!');
+        });
+    }
+    else {
+      alert('You do not have permission to perform this action.');
+    }
   }
 }
